third_party/WebKit/LayoutTests/css1/units/rounding-expected.txt

no gap below
no gap above

This test checks that floating point rounding doesn't cause misalignment. There should be no gap between the divs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Math.round(divtop.bottom) is 31
PASS Math.round(divbottom.top) is 31
PASS successfullyParsed is true

TEST COMPLETE
